--- Metalord_btin/MetaLord/Assets/_Test/KHJ/Models/Node-Mesh.mesh.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 2a57c1928931caf4d94d77688e351c33
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 4300000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
