+++ Metalord_btin/MetaLord/Assets/_Test/KHJ/Prefabs/ImportantPrefab/Npc/KitchenRoomNpc/DinosaurNpc.prefab
@@ -1,5 +1,73 @@
 %YAML 1.1
 %TAG !u! tag:unity3d.com,2011:
+--- !u!1 &2837547637132976099
+GameObject:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  serializedVersion: 6
+  m_Component:
+  - component: {fileID: 1287640484358300661}
+  - component: {fileID: 7251500478514117235}
+  - component: {fileID: 4297138291546391816}
+  m_Layer: 13
+  m_Name: InteractZone
+  m_TagString: Untagged
+  m_Icon: {fileID: 0}
+  m_NavMeshLayer: 0
+  m_StaticEditorFlags: 0
+  m_IsActive: 1
+--- !u!4 &1287640484358300661
+Transform:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 2837547637132976099}
+  serializedVersion: 2
+  m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
+  m_LocalPosition: {x: 0, y: 0, z: 0}
+  m_LocalScale: {x: 1, y: 1, z: 1}
+  m_ConstrainProportionsScale: 0
+  m_Children: []
+  m_Father: {fileID: 6017562838789853775}
+  m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
+--- !u!65 &7251500478514117235
+BoxCollider:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 2837547637132976099}
+  m_Material: {fileID: 0}
+  m_IncludeLayers:
+    serializedVersion: 2
+    m_Bits: 0
+  m_ExcludeLayers:
+    serializedVersion: 2
+    m_Bits: 0
+  m_LayerOverridePriority: 0
+  m_IsTrigger: 1
+  m_ProvidesContacts: 0
+  m_Enabled: 1
+  serializedVersion: 3
+  m_Size: {x: 0.42, y: 0.47012353, z: 0.36}
+  m_Center: {x: 0.11, y: 0.26, z: 0.38}
+--- !u!114 &4297138291546391816
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 2837547637132976099}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: cb5c3a6d9dca6f34e87c96a478dcd13c, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  fadeImage: {fileID: 3673953700429976785}
+  isShowInteractUI: 0
 --- !u!1 &2996097611897435313
 GameObject:
   m_ObjectHideFlags: 0
@@ -116,7 +184,6 @@ GameObject:
   - component: {fileID: 310373173625041327}
   - component: {fileID: 141647730312291726}
   - component: {fileID: 8154570452828040289}
-  - component: {fileID: 4138404602854813557}
   - component: {fileID: 9033758854217198815}
   m_Layer: 8
   m_Name: DinosaurNpc
@@ -139,6 +206,7 @@ Transform:
   m_ConstrainProportionsScale: 0
   m_Children:
   - {fileID: 2315403152343273104}
+  - {fileID: 1287640484358300661}
   m_Father: {fileID: 0}
   m_LocalEulerAnglesHint: {x: 0, y: 160.925, z: 0}
 --- !u!33 &4060028195662379325
@@ -210,8 +278,8 @@ BoxCollider:
   m_ProvidesContacts: 0
   m_Enabled: 1
   serializedVersion: 3
-  m_Size: {x: 0.42, y: 0.47012353, z: 0.36}
-  m_Center: {x: 0.11, y: 0.26, z: 0.38}
+  m_Size: {x: 0.1, y: 0.89, z: 0.1}
+  m_Center: {x: 0, y: -200, z: 0.38}
 --- !u!64 &141647730312291726
 MeshCollider:
   m_ObjectHideFlags: 0
@@ -247,19 +315,6 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   myDialogue: {fileID: 0}
---- !u!114 &4138404602854813557
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 4107117629837978318}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: cb5c3a6d9dca6f34e87c96a478dcd13c, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  fadeImage: {fileID: 3673953700429976785}
 --- !u!114 &9033758854217198815
 MonoBehaviour:
   m_ObjectHideFlags: 0
