+++ Metalord_btin/MetaLord/Assets/_Test/KHJ/Prefabs/ImportantPrefab/Npc/KitchenRoomNpc/SheepNpc.prefab
@@ -14,7 +14,6 @@ GameObject:
   - component: {fileID: 6773046664397870340}
   - component: {fileID: 692808885318871486}
   - component: {fileID: 8337649906800620282}
-  - component: {fileID: 1580330006859474208}
   - component: {fileID: 410666043207146335}
   m_Layer: 8
   m_Name: SheepNpc
@@ -37,6 +36,7 @@ Transform:
   m_ConstrainProportionsScale: 0
   m_Children:
   - {fileID: 6518619193451061525}
+  - {fileID: 7116670343199945898}
   m_Father: {fileID: 0}
   m_LocalEulerAnglesHint: {x: 0, y: 42.203, z: 0}
 --- !u!33 &5985997623150112681
@@ -108,8 +108,8 @@ BoxCollider:
   m_ProvidesContacts: 0
   m_Enabled: 1
   serializedVersion: 3
-  m_Size: {x: 0.31852397, y: 0.43406025, z: 0.4727531}
-  m_Center: {x: 0.31, y: 0.051867798, z: -0.24}
+  m_Size: {x: 0.001, y: 0.001, z: 0.001}
+  m_Center: {x: 0.31, y: -200, z: -0.24}
 --- !u!64 &692808885318871486
 MeshCollider:
   m_ObjectHideFlags: 0
@@ -145,19 +145,6 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   myDialogue: {fileID: 0}
---- !u!114 &1580330006859474208
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 477451153861566170}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: cb5c3a6d9dca6f34e87c96a478dcd13c, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  fadeImage: {fileID: 1703336173643174773}
 --- !u!114 &410666043207146335
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -251,6 +238,74 @@ MonoBehaviour:
   m_FillOrigin: 0
   m_UseSpriteMesh: 0
   m_PixelsPerUnitMultiplier: 1
+--- !u!1 &8539231348923309317
+GameObject:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  serializedVersion: 6
+  m_Component:
+  - component: {fileID: 7116670343199945898}
+  - component: {fileID: 5768060641163601830}
+  - component: {fileID: 5666196890282890583}
+  m_Layer: 13
+  m_Name: InteractZone
+  m_TagString: Untagged
+  m_Icon: {fileID: 0}
+  m_NavMeshLayer: 0
+  m_StaticEditorFlags: 0
+  m_IsActive: 1
+--- !u!4 &7116670343199945898
+Transform:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 8539231348923309317}
+  serializedVersion: 2
+  m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
+  m_LocalPosition: {x: 0, y: 0, z: 0}
+  m_LocalScale: {x: 1, y: 1, z: 1}
+  m_ConstrainProportionsScale: 0
+  m_Children: []
+  m_Father: {fileID: 4208122371503013546}
+  m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
+--- !u!114 &5768060641163601830
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 8539231348923309317}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: cb5c3a6d9dca6f34e87c96a478dcd13c, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  fadeImage: {fileID: 1703336173643174773}
+  isShowInteractUI: 0
+--- !u!65 &5666196890282890583
+BoxCollider:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 8539231348923309317}
+  m_Material: {fileID: 0}
+  m_IncludeLayers:
+    serializedVersion: 2
+    m_Bits: 0
+  m_ExcludeLayers:
+    serializedVersion: 2
+    m_Bits: 0
+  m_LayerOverridePriority: 0
+  m_IsTrigger: 1
+  m_ProvidesContacts: 0
+  m_Enabled: 1
+  serializedVersion: 3
+  m_Size: {x: 0.31852397, y: 0.43406025, z: 0.4727531}
+  m_Center: {x: 0.31, y: 0.051867798, z: -0.24}
 --- !u!1 &8619147687101516677
 GameObject:
   m_ObjectHideFlags: 0
