+++ Metalord_btin/MetaLord/Assets/_Test/KHJ/Prefabs/ImportantPrefab/Npc/MiddleRoomNpc/NarwhalNpc.prefab
@@ -14,7 +14,6 @@ GameObject:
   - component: {fileID: 3515185507321058585}
   - component: {fileID: 7528997530552389270}
   - component: {fileID: 7753100664843252237}
-  - component: {fileID: 2889247587253340407}
   - component: {fileID: 5727004788947952254}
   m_Layer: 8
   m_Name: NarwhalNpc
@@ -37,6 +36,7 @@ Transform:
   m_ConstrainProportionsScale: 0
   m_Children:
   - {fileID: 2149500076786244051}
+  - {fileID: 3265808055197178913}
   m_Father: {fileID: 0}
   m_LocalEulerAnglesHint: {x: 0, y: -7.96, z: 0}
 --- !u!33 &2980936460022371283
@@ -121,8 +121,8 @@ BoxCollider:
   m_ProvidesContacts: 0
   m_Enabled: 1
   serializedVersion: 3
-  m_Size: {x: 0.43, y: 0.30737978, z: 0.48}
-  m_Center: {x: 9.476439e-10, y: 0.14940831, z: 0.45}
+  m_Size: {x: 0, y: 0.0001, z: 0}
+  m_Center: {x: 0, y: -200, z: 0}
 --- !u!64 &7753100664843252237
 MeshCollider:
   m_ObjectHideFlags: 0
@@ -145,19 +145,6 @@ MeshCollider:
   m_Convex: 0
   m_CookingOptions: 30
   m_Mesh: {fileID: -223148573040256356, guid: 732ede9455a7780438afab98b228a94d, type: 3}
---- !u!114 &2889247587253340407
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 868962798350387523}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: cb5c3a6d9dca6f34e87c96a478dcd13c, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  fadeImage: {fileID: 4916088353210781641}
 --- !u!114 &5727004788947952254
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -278,6 +265,74 @@ MonoBehaviour:
   m_BlockingMask:
     serializedVersion: 2
     m_Bits: 4294967295
+--- !u!1 &2574909618575722514
+GameObject:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  serializedVersion: 6
+  m_Component:
+  - component: {fileID: 3265808055197178913}
+  - component: {fileID: 8552671044150580764}
+  - component: {fileID: 1578893112002547299}
+  m_Layer: 13
+  m_Name: InteractZone
+  m_TagString: Untagged
+  m_Icon: {fileID: 0}
+  m_NavMeshLayer: 0
+  m_StaticEditorFlags: 0
+  m_IsActive: 1
+--- !u!4 &3265808055197178913
+Transform:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 2574909618575722514}
+  serializedVersion: 2
+  m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
+  m_LocalPosition: {x: 0, y: 0, z: 0}
+  m_LocalScale: {x: 1, y: 1, z: 1}
+  m_ConstrainProportionsScale: 0
+  m_Children: []
+  m_Father: {fileID: 6499929154988556345}
+  m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
+--- !u!65 &8552671044150580764
+BoxCollider:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 2574909618575722514}
+  m_Material: {fileID: 0}
+  m_IncludeLayers:
+    serializedVersion: 2
+    m_Bits: 0
+  m_ExcludeLayers:
+    serializedVersion: 2
+    m_Bits: 0
+  m_LayerOverridePriority: 0
+  m_IsTrigger: 1
+  m_ProvidesContacts: 0
+  m_Enabled: 1
+  serializedVersion: 3
+  m_Size: {x: 0.43, y: 0.30737978, z: 0.48}
+  m_Center: {x: 9.476439e-10, y: 0.14940831, z: 0.45}
+--- !u!114 &1578893112002547299
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 2574909618575722514}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: cb5c3a6d9dca6f34e87c96a478dcd13c, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  fadeImage: {fileID: 4916088353210781641}
+  isShowInteractUI: 0
 --- !u!1 &6736932836465005594
 GameObject:
   m_ObjectHideFlags: 0
