+++ Metalord_btin/MetaLord/Assets/_Test/KHJ/Scripts/InteractableNpc/AdventuerNpc.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 0d0d1618260f6f844a8657bbd383866a
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
