--- Metalord_btin/MetaLord/Assets/_Test/KHJ/Scripts/InteractableNpc/AdventurerNpc.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 3dd29179a5b6be745a43928134b4a033
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
