--- Metalord_btin/MetaLord/Assets/_Test/KHJ/Scripts/InteractableNpc/DinosaurNpc.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 5b410cdcdc8b6bb4bb5be11d0137de76
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
