+++ Metalord_btin/MetaLord/Assets/_Test/KHJ/Scripts/InteractableNpc/DinosaurNpc.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 85a7cddbe08062647ae00abf50d32c0f
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
