+++ Metalord_btin/MetaLord/Assets/_Test/KHJ/Scripts/InteractableNpc/DinosaurNpc/DinosaurNpc.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 5b410cdcdc8b6bb4bb5be11d0137de76
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
