+++ Metalord_btin/MetaLord/Assets/_Test/PSC/Materials/NonPaintable/Toys_TextureAtlas_01_standard.mat
@@ -21,7 +21,7 @@ Material:
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
   m_Name: Toys_TextureAtlas_01_standard
-  m_Shader: {fileID: 46, guid: 0000000000000000f000000000000000, type: 0}
+  m_Shader: {fileID: 4800000, guid: 2b0f5fce771a1c645adeaed55998c8f1, type: 3}
   m_Parent: {fileID: 0}
   m_ModifiedSerializedProperties: 0
   m_ValidKeywords: []
