+++ Metalord_btin/MetaLord/ProjectSettings/TagManager.asset
@@ -18,7 +18,7 @@ TagManager:
   - MovedObject
   - CatchObject
   - NonInteractObject
-  - 
+  - InteractZone
   - 
   - 
   -
